Move _cached_angular_distribution and make_log_interpolator to utils
